Reorg scatter integration code into maintainable and scalable file structure
